backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
requests==2.31.0
openai==1.3.7
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
sqlalchemy==2.0.23
alembic==1.13.1
email-validator==2.1.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3
# Authentication System Requirements for Agent-Cofounder

# Core Framework
flask>=2.3.0
werkzeug>=2.3.0

# Supabase Integration
supabase>=2.0.0

# Data Validation
pydantic>=2.0.0
email-validator>=2.0.0

# Date Handling
python-dateutil>=2.8.0

# Environment Management (optional)
python-dotenv>=1.0.0